Solved issues: 1. There are two clients for the preprocessor so the session ids for pre and post are different - can not get one job_uid - share the same websocket client if it is the same server address? Will not quite work as even if they are different one might want to use the source file for post processing on the server side. 2. The job id of the processor request is issued by the static manager so it is not the same as there are two pre and post managers - so just have one fixed job id = 1 for both jobs - in the command line client.
The job token is now issued by the client and updated by the server. If the session is canceled its files are not deleted. Only a finished processor job cleans after itself. Thus there can be some files remained over but this is only in case of error situations, these can be cleaned up by some cron jobs